requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
Brotli==1.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
ffmpeg-python==0.2.0
frozenlist==1.7.0
future==1.0.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.3
mutagen==1.47.0
propcache==0.3.2
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
pytube==15.0.0
requests==2.32.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
yt-dlp==2025.6.30
